requirements.txt
torch
torchvision
scikit-learn
scikit-image
aiogram
python-dotenv
matplotlib
pandas
joblib
tqdm
Pillow